Fix props names in EditButton and PetForm, moved pet-related components to components/pages/pets
