+++ vet-spring/src/pages/pets/EditButton.jsx
@@ -1,8 +1,8 @@
 import { useContext } from "react";
-import ThemeContext from "../utils/helpers/themeContext";
-import AddPetForm from "./PetForm";
+import ThemeContext from "../../utils/helpers/themeContext";
+import PetForm from "./PetForm";
 
-function EditButton({ pet }) {
+function EditButton({ pet, getPage, currentPage, pageSize }) {
     const { editModalID, setEditModalID } = useContext(ThemeContext);
 
     return (
@@ -21,7 +21,7 @@ function EditButton({ pet }) {
             {editModalID == pet.id && (
                 <dialog open className="modal">
                     <div className="modal-box bg-gray-600 text-center">
-                        <AddPetForm pet={pet} getPetPage={getPage} currentPage={currentPage} pageSize={pageSize} />
+                        <PetForm pet={pet} getPage={getPage} currentPage={currentPage} pageSize={pageSize} />
                         <form method="dialog">
                             <button
                                 className="btn bg-red-500 w-16"
