+++ vet-spring/src/pages/pets/PetCard.jsx
@@ -1,5 +1,5 @@
 import DeleteButton from "../../components/DeleteButton"
-import EditButton from "../../components/EditButton"
+import EditButton from "./EditButton"
 
 export const PetCard = ({ pet, getPage, currentPage, pageSize }) => {
     const { name, species, breed, birthdate, gender } = pet;
@@ -15,7 +15,7 @@ export const PetCard = ({ pet, getPage, currentPage, pageSize }) => {
                 <p className="py-2">{gender}</p>
             </div>
             <div className="card-actions">
-                <EditButton pet={pet} />
+                <EditButton pet={pet} getPage={getPage} currentPage={currentPage} pageSize={pageSize}/>
                 <DeleteButton pet={pet} getPage={getPage} currentPage={currentPage} pageSize={pageSize} />
             </div>
         </div>
