+++ vet-spring/src/pages/pets/PetForm.jsx
@@ -1,8 +1,8 @@
 import { useEffect, useState } from "react";
 import { useForm } from "react-hook-form";
-import { addPet, updatePet } from "../utils/helpers/petService";
+import { addPet, updatePet } from "../../utils/helpers/petService";
 
-const AddPetForm = ({ pet, getPetPage, currentPage, pageSize }) => {
+const PetForm = ({ pet, getPage, currentPage, pageSize }) => {
   const {
     register,
     handleSubmit,
@@ -33,6 +33,8 @@ const AddPetForm = ({ pet, getPetPage, currentPage, pageSize }) => {
     const trimmedData = {
       ...data,
       name: data.name.trim(),
+      species: data.species.trim(),
+      breed: data.breed.trim(),
       birthdate: data.birthdate || null,
     }
 
@@ -41,11 +43,11 @@ const AddPetForm = ({ pet, getPetPage, currentPage, pageSize }) => {
     try {
       if (pet && pet.id) {
         await updatePet(pet.id, payload);
-        await getPetPage(pageSize, currentPage);
+        await getPage(pageSize, currentPage);
       } else {
         const newPayload = { ...trimmedData };
         await addPet(newPayload);
-        await getPetPage(pageSize, currentPage);
+        await getPage(pageSize, currentPage); 
       }
       reset({ name: "", species: "", breed: "", birthdate: "", gender: "" });
     } catch (error) {
@@ -163,4 +165,4 @@ const AddPetForm = ({ pet, getPetPage, currentPage, pageSize }) => {
   );
 };
 
-export default AddPetForm;
+export default PetForm;
